# appBackend/share/openapi.yaml
openapi: 3.0.0
info:
  title: 'Timesheet backend application'
  description: 'Api Timesheet backend application'
  version: 1.0.0
paths:
...
  /api/timesheet/tickets:
    get:
      tags:
        - Timesheet
      summary: 'Получить список тикетов'
      description: 'Возвращает список тикетов учёта времени с фильтрацией и пагинацией.'
      operationId: get_timesheet_list_tickets
      parameters:
        -
          name: employeeId
          in: query
          description: 'Фильтр по сотруднику.'
          required: false
          schema:
            type: string
            format: uuid
        -
          name: projectId
          in: query
          description: 'Фильтр по проекту.'
          required: false
          schema:
            type: string
            format: uuid
        -
          name: crId
          in: query
          description: 'Фильтр по запросу на изменение.'
          required: false
          schema:
            type: string
            format: uuid
        -
          name: taskId
          in: query
          description: 'Фильтр по задаче.'
          required: false
          schema:
            type: string
            format: uuid
        -
          name: workId
          in: query
          description: 'Фильтр по виду работ.'
          required: false
          schema:
            type: string
            format: uuid
        -
          name: dateFrom
          in: query
          description: 'Начальная дата фильтра.'
          required: false
          schema:
            type: string
            format: date
        -
          name: dateTo
          in: query
          description: 'Конечная дата фильтра.'
          required: false
          schema:
            type: string
            format: date
        -
          name: page
          in: query
          description: 'Номер страницы.'
          required: false
          schema:
            type: integer
            default: 1
        -
          name: perPage
          in: query
          description: 'Количество элементов на странице.'
          required: false
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 'Список тикетов.'
          content:
            application/json:
              schema:
                $ref: App\Timesheet\Infrastructure\Dto\TicketListResponseDto
    post:
      tags:
        - Timesheet
      summary: 'Создать тикет'
      description: 'Создаёт новый тикет учёта времени.'
      operationId: post_timesheet_create_ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: App\Timesheet\Infrastructure\Dto\CreateManualTicketRequestDto
      responses:
        '201':
          description: 'Тикет успешно создан.'
        '400':
          description: 'Невалидные данные.'
        '403':
          description: 'Нарушение прав доступа.'
        '404':
          description: 'Задача или вид работ не найдены.'
        '422':
          description: 'Нарушение бизнес-правил.'
...
  /api/work-catalog/rates:
    get:
      tags:
        - WorkCatalog
      summary: 'Получить список ставок'
      description: 'Возвращает список ставок с пагинацией и фильтрацией.'
      operationId: get_work_catalog_list_rates
      parameters:
        -
          name: roleId
          in: query
          description: 'Фильтр по роли.'
          required: false
          schema:
            type: string
            format: uuid
        -
          name: workId
          in: query
          description: 'Фильтр по виду работ.'
          required: false
          schema:
            type: string
            format: uuid
        -
          name: page
          in: query
          description: 'Номер страницы.'
          required: false
          schema:
            type: integer
            default: 1
        -
          name: perPage
          in: query
          description: 'Количество элементов на странице.'
          required: false
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 'Список ставок.'
          content:
            application/json:
              schema:
                $ref: App\WorkCatalog\Infrastructure\Dto\RateListResponseDto
    post:
      tags:
        - WorkCatalog
      summary: 'Создать ставку'
      description: 'Создаёт новую ставку.'
      operationId: post_work_catalog_create_rate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: App\WorkCatalog\Infrastructure\Dto\CreateRateRequestDto
      responses:
        '201':
          description: 'Ставка успешно создана.'
        '400':
          description: 'Невалидные данные запроса.'
        '404':
          description: 'Роль или вид работ не найдены.'
        '422':
          description: 'Нарушение бизнес-правил.'
...
  '/api/work-catalog/rates/{id}':
    put:
      tags:
        - WorkCatalog
      summary: 'Обновить ставку'
      description: 'Обновляет данные существующей ставки.'
      operationId: put_work_catalog_update_rate
      parameters:
        -
          name: id
          in: path
          description: 'Идентификатор ставки.'
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: App\WorkCatalog\Infrastructure\Dto\UpdateRateRequestDto
      responses:
        '200':
          description: 'Ставка успешно обновлена.'
        '404':
          description: 'Ставка, роль или вид работ не найдены.'
        '422':
          description: 'Нарушение бизнес-правил.'
    delete:
      tags:
        - WorkCatalog
      summary: 'Удалить ставку'
      description: 'Выполняет мягкое удаление ставки.'
      operationId: delete_work_catalog_delete_rate
      parameters:
        -
          name: id
          in: path
          description: 'Идентификатор ставки.'
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: 'Ставка успешно удалена.'
        '404':
          description: 'Ставка не найдена.'
        '422':
          description: 'Нарушение бизнес-правил.'
...